formsAdvanced/requirements.txt
asgiref==3.11.0
crispy-bootstrap5==2025.6
Django==5.2.8
django-crispy-forms==2.5
Markdown==3.10
pi==0.1.2
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
